requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
openai==1.3.0
python-multipart==0.0.6
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-filter==23.5
